Add venv for maturin in CI, exclude legacy tests missing ansible_vault
